# apps/web/src/lib/service-clients/service-auth/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "authentication_service",
    "description": "",
    "termsOfService": "https://macro.com/terms",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/login/passwordless": {
      "post": {
        "tags": ["login::passwordless"],
        "summary": "Initiates a passwordless login",
        "operationId": "passwordless_login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordlessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordlessStartedResponse"
                }
              }
            }
          },
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SsoRequiredResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/user": {
      "post": {
        "tags": ["user::create_user"],
        "summary": "Creates a new user.",
        "operationId": "create_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateUserRequest": {
        "type": "object",
        "description": "The request body to create a new user in fusionauth\nNOTE: Never derive debug here as we don't want to accidentally log the password",
        "required": ["username", "email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "description": "The primary email address of the user.\nThis will be the user's root \"profile\"."
          },
          "password": {
            "type": "string",
            "description": "The password for the user.\nTODO: configure password policy and validate password before attempting to create user"
          },
          "username": {
            "type": "string",
            "description": "The unique username for the user."
          }
        }
      },
      "EmptyResponse": {
        "type": "object",
        "description": "Empty response is required due to custom fetch forcing `response.json()`"
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "A plain old json error response for use with axum.\nyup, thats it.",
        "required": ["message"],
        "properties": {
          "message": {
            "type": "string",
            "description": "Message to explain failure"
          }
        }
      },
...
      "PasswordlessRequest": {
        "type": "object",
        "required": ["email", "redirect_uri"],
        "properties": {
          "email": {
            "type": "string",
            "description": "The email to initiate passwordless authentication for"
          },
          "redirect_uri": {
            "type": "string",
            "description": "The redirect uri to redirect to after login"
          },
          "referral_code": {
            "type": ["string", "null"],
            "description": "The referral code"
          }
        }
      },
      "PasswordlessStartedResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": ["string", "null"],
            "description": "The one-time login code, returned only by local environments so dev\ntooling can complete the flow without reading the email."
          }
        }
      },
...
      "SsoRequiredResponse": {
        "type": "object",
        "required": ["idp_id"],
        "properties": {
          "idp_id": {
            "type": "string",
            "description": "The idp_id you need to perform sso login with."
          }
        }
      },
...
    }
  },
...
}